Remove QgsDataDefinedButton
Now completely replaced by QgsPropertyOverrideButton
