+++ src/core/qgsexpressioncontextgenerator.h
@@ -25,7 +25,7 @@
  * You need to implement this interface in a class and register this class with
  * QgsFieldExpressionWidget::registerExpressionGenerator().
  *
- * This is used for example in QgsDataDefinedButton or QgsFieldExpressionWidget
+ * This is used for example in QgsPropertyOverrideButton or QgsFieldExpressionWidget
  * classes which will ask for a new QgsExpressionContext every time the expression
  * editor is opened. This way they are able to provide an up-to-date expression
  * editor even when the environment changes.
